+++ src/main/java/com/so/book/admin/review/AdReviewMapper.java
@@ -14,9 +14,9 @@ public interface AdReviewMapper {
 
 	int review_count(@Param("cri") SearchCriteria cri, @Param("rev_content") String rev_content, @Param("rev_score") String rev_score);
 	
-	ReviewReply reply_info(Long reply_id);
+	ReviewReply reply_info(Integer reply_id);
 	
-	void reply_modify_save(@Param("reply_id") Long reply_id, @Param("reply_content") String reply_content);
+	void reply_modify_save(@Param("reply_id") Integer reply_id, @Param("reply_content") String reply_content);
 	
-	void reply_delete(Long reply_id);
+	void reply_delete(Integer reply_id);
 }
