+++ src/main/java/com/so/book/review/ReviewVo.java
@@ -14,7 +14,7 @@ import lombok.ToString;
 @ToString
 public class ReviewVo {
 
-	private Long rev_code;
+	private Integer rev_code;
 	private String mem_id;
 	private Integer pro_code;
 	private String rev_title;
